Moves SupportingText e2e tests to their own file
